+++ packages/router/src/bindings/contractReader/index.ts
@@ -53,13 +53,15 @@ export const bindContractReader = async () => {
         logger.info("active and handling tracker", requestContext, methodContext, {
           transactionsLength: transactions.length,
           handlingTrackerLength: handlingTracker.size,
+          byStatus: Object.keys(CrosschainTransactionStatus).map((status) => {
+            return { status, size: [...handlingTracker.values()].map((v) => v.status === status).length };
+          }),
         });
         logger.debug("active and handling tracker details", requestContext, methodContext, {
-          transactions: transactions,
           handlingTracker: [...handlingTracker],
         });
       }
-    } catch (err) {
+    } catch (err: any) {
       logger.error("Error getting active txs, waiting for next loop", requestContext, methodContext, jsonifyError(err));
       return;
     }
@@ -179,16 +181,22 @@ export const handleSingle = async (
         requestContext,
       });
     }
+    if (!_transaction.payload.hashes.sending) {
+      logger.warn("No sending hashes with SenderPrepared status", requestContext, methodContext, {
+        hashes: _transaction.payload.hashes,
+      });
+      return;
+    }
     const senderPrepareReceipt = await txService.getTransactionReceipt(
       _transaction.crosschainTx.invariant.sendingChainId,
-      _transaction.payload.senderPreparedHash,
+      _transaction.payload.hashes.sending.prepareHash,
     );
     if ((senderPrepareReceipt?.confirmations ?? 0) < chainConfig.confirmations) {
       logger.info("Waiting for safe confirmations", requestContext, methodContext, {
         txConfirmations: senderPrepareReceipt?.confirmations ?? 0,
         configuredConfirmations: chainConfig.confirmations,
         chainId: _transaction.crosschainTx.invariant.sendingChainId,
-        txHash: _transaction.payload.senderPreparedHash,
+        txHash: _transaction.payload.hashes.sending.prepareHash,
       });
       return;
     }
@@ -222,7 +230,7 @@ export const handleSingle = async (
         chainId: _transaction.crosschainTx.invariant.receivingChainId,
         amount: receipt!.gasUsed.toString(),
       });
-    } catch (err) {
+    } catch (err: any) {
       const json = jsonifyError(err);
       if (safeJsonStringify(json).includes("#P:015")) {
         logger.warn("Receiver transaction already prepared", requestContext, methodContext, { error: json });
@@ -245,6 +253,7 @@ export const handleSingle = async (
               amount: transaction.crosschainTx.sending.amount,
               expiry: transaction.crosschainTx.sending.expiry,
               preparedBlockNumber: transaction.crosschainTx.sending.preparedBlockNumber,
+              preparedTransactionHash: transaction.payload.hashes.sending.prepareHash,
               side: "sender",
             },
             requestContext,
@@ -259,7 +268,7 @@ export const handleSingle = async (
             chainId: _transaction.crosschainTx.invariant.sendingChainId,
             amount: cancelRes!.gasUsed.toString(),
           });
-        } catch (cancelErr) {
+        } catch (cancelErr: any) {
           const cancelJson = jsonifyError(cancelErr);
           if (safeJsonStringify(jsonifyError(cancelErr)).includes("#C:019")) {
             logger.warn("Already cancelled", requestContext, methodContext, {
@@ -285,14 +294,20 @@ export const handleSingle = async (
       // this should not happen, this should get checked before this point
       throw new ContractReaderNotAvailableForChain(_transaction.crosschainTx.invariant.sendingChainId, {});
     }
+    if (!_transaction.payload.hashes.receiving) {
+      logger.warn("No receiving hashes with ReceiverFulfilled status", requestContext, methodContext, {
+        hashes: _transaction.payload.hashes,
+      });
+      return;
+    }
     const receiverPrepareReceipt = await txService.getTransactionReceipt(
       _transaction.crosschainTx.invariant.receivingChainId,
-      _transaction.payload.receiverFulfilledHash,
+      _transaction.payload.hashes.receiving.prepareHash,
     );
     if ((receiverPrepareReceipt?.confirmations ?? 0) < chainConfig.confirmations) {
       logger.info("Waiting for safe confirmations", requestContext, methodContext, {
         chainId: _transaction.crosschainTx.invariant.receivingChainId,
-        txHash: _transaction.payload.receiverFulfilledHash,
+        txHash: _transaction.payload.hashes.receiving.prepareHash,
         txConfirmations: receiverPrepareReceipt?.confirmations ?? 0,
         configuredConfirmations: chainConfig.confirmations,
       });
@@ -344,7 +359,7 @@ export const handleSingle = async (
         chainId: _transaction.crosschainTx.invariant.sendingChainId,
         amount: receipt!.gasUsed.toString(),
       });
-    } catch (err) {
+    } catch (err: any) {
       const jsonErr = jsonifyError(err);
       if (safeJsonStringify(jsonErr).includes("#F:019")) {
         logger.warn("Sender already fulfilled", requestContext, methodContext, { error: jsonErr });
@@ -364,6 +379,12 @@ export const handleSingle = async (
       transaction.crosschainTx.invariant.transactionId,
     );
     const _transaction = transaction as ActiveTransaction<"ReceiverExpired">;
+    if (!_transaction.payload.hashes.receiving) {
+      logger.warn("No receiving hashes with ReceiverExpired status", requestContext, methodContext, {
+        hashes: _transaction.payload.hashes,
+      });
+      return;
+    }
     try {
       logger.info("Cancelling expired receiver", requestContext, methodContext);
       receipt = await cancel(
@@ -372,6 +393,7 @@ export const handleSingle = async (
           amount: _transaction.crosschainTx.receiving!.amount,
           expiry: _transaction.crosschainTx.receiving!.expiry,
           preparedBlockNumber: _transaction.crosschainTx.receiving!.preparedBlockNumber,
+          preparedTransactionHash: _transaction.payload.hashes.receiving.prepareHash,
           side: "receiver",
         },
         requestContext,
@@ -388,7 +410,7 @@ export const handleSingle = async (
         chainId: _transaction.crosschainTx.invariant.receivingChainId,
         amount: receipt!.gasUsed.toString(),
       });
-    } catch (err) {
+    } catch (err: any) {
       const errJson = jsonifyError(err);
       if (safeJsonStringify(errJson).includes("#C:019")) {
         logger.warn("Already cancelled", requestContext, methodContext, { error: errJson });
@@ -408,6 +430,12 @@ export const handleSingle = async (
       transaction.crosschainTx.invariant.transactionId,
     );
     const _transaction = transaction as ActiveTransaction<"SenderExpired">;
+    if (!_transaction.payload.hashes.sending) {
+      logger.warn("No sending hashes with SenderExpired status", requestContext, methodContext, {
+        hashes: _transaction.payload.hashes,
+      });
+      return;
+    }
     try {
       logger.info("Cancelling expired sender", requestContext, methodContext);
       receipt = await cancel(
@@ -416,6 +444,7 @@ export const handleSingle = async (
           amount: _transaction.crosschainTx.sending.amount,
           expiry: _transaction.crosschainTx.sending.expiry,
           preparedBlockNumber: _transaction.crosschainTx.sending.preparedBlockNumber,
+          preparedTransactionHash: _transaction.payload.hashes.sending.prepareHash,
           side: "sender",
         },
         requestContext,
@@ -430,7 +459,7 @@ export const handleSingle = async (
         chainId: _transaction.crosschainTx.invariant.sendingChainId,
         amount: receipt!.gasUsed.toString(),
       });
-    } catch (err) {
+    } catch (err: any) {
       const errJson = jsonifyError(err);
       if (safeJsonStringify(errJson).includes("#C:019")) {
         logger.warn("Already cancelled", requestContext, methodContext, { error: errJson });
@@ -454,6 +483,12 @@ export const handleSingle = async (
       "ContractReader => ReceiverCancelled",
       transaction.crosschainTx.invariant.transactionId,
     );
+    if (!_transaction.payload.hashes.sending) {
+      logger.warn("No sending hashes with ReceiverCancelled status", requestContext, methodContext, {
+        hashes: _transaction.payload.hashes,
+      });
+      return;
+    }
     try {
       logger.info("Cancelling sender after receiver cancelled", requestContext, methodContext);
       receipt = await cancel(
@@ -462,6 +497,7 @@ export const handleSingle = async (
           amount: _transaction.crosschainTx.sending.amount,
           expiry: _transaction.crosschainTx.sending.expiry,
           preparedBlockNumber: _transaction.crosschainTx.sending.preparedBlockNumber,
+          preparedTransactionHash: _transaction.payload.hashes.sending.prepareHash,
           side: "sender",
         },
         requestContext,
@@ -476,7 +512,7 @@ export const handleSingle = async (
         chainId: _transaction.crosschainTx.invariant.sendingChainId,
         amount: receipt!.gasUsed.toString(),
       });
-    } catch (err) {
+    } catch (err: any) {
       const errJson = jsonifyError(err);
       if (safeJsonStringify(errJson).includes("#C:019")) {
         logger.warn("Already cancelled", requestContext, methodContext, { error: errJson });
@@ -497,6 +533,12 @@ export const handleSingle = async (
       "ContractReader => ReceiverNotConfigured",
       transaction.crosschainTx.invariant.transactionId,
     );
+    if (!_transaction.payload.hashes.sending) {
+      logger.warn("No sending hashes with ReceiverNotConfigured status", requestContext, methodContext, {
+        hashes: _transaction.payload.hashes,
+      });
+      return;
+    }
     try {
       logger.info("Cancelling sender because receiver is not configured", requestContext, methodContext);
       receipt = await cancel(
@@ -505,6 +547,7 @@ export const handleSingle = async (
           amount: _transaction.crosschainTx.sending.amount,
           expiry: _transaction.crosschainTx.sending.expiry,
           preparedBlockNumber: _transaction.crosschainTx.sending.preparedBlockNumber,
+          preparedTransactionHash: _transaction.payload.hashes.sending.prepareHash,
           side: "sender",
         },
         requestContext,
@@ -519,7 +562,7 @@ export const handleSingle = async (
         chainId: _transaction.crosschainTx.invariant.sendingChainId,
         amount: receipt!.gasUsed.toString(),
       });
-    } catch (err) {
+    } catch (err: any) {
       const errJson = jsonifyError(err);
       if (safeJsonStringify(errJson).includes("#C:019")) {
         logger.warn("Already cancelled", requestContext, methodContext, { error: errJson });