+++ packages/router/src/config.ts
@@ -124,6 +124,7 @@ export const TSwapPool = Type.Object({
       assetId: TAddress,
     }),
   ),
+  mainnetEquivalent: Type.Optional(TAddress),
 });
 
 export const NxtpRouterConfigSchema = Type.Object({