+++ packages/router/src/lib/operations/prepare.ts
@@ -16,6 +16,7 @@ import {
   NotEnoughLiquidity,
   SenderChainDataInvalid,
   BidExpiryInvalid,
+  NotEnoughAmount,
 } from "../errors";
 import {
   decodeAuctionBid,
@@ -178,6 +179,17 @@ export const prepare = async (
         subgraphBalance: routerBalance.toString(),
       });
     }
+  }
+
+  // Make sure amount is sensible
+  if (BigNumber.from(receiverAmount).lt(0)) {
+    throw new NotEnoughAmount({
+      receiverAmount,
+      preparedAmount: senderAmount.toString(),
+      transactionId: invariantData.transactionId,
+      methodContext,
+      requestContext,
+    });
   }
 
   // Handle the expiries.