+++ packages/router/test/bindings/prices/index.spec.ts
@@ -1,4 +1,4 @@
-import { SinonStub, stub } from "sinon";
+import { reset, restore, SinonStub, stub } from "sinon";
 import { parseEther } from "@ethersproject/units";
 import { delay, expect, mkAddress } from "@connext/nxtp-utils";
 
@@ -30,36 +30,34 @@ describe("Fetching Price Binding", () => {
       getTokenPriceFromOnChainStub = stub(SharedFns, "getTokenPriceFromOnChain");
     });
 
+    afterEach(() => {
+      restore();
+      reset();
+    });
+
     it("should error if multicall is failed", async () => {
       multicallStub.throws(new Error("fail"));
       await binding.bindPrices();
       await delay(interval + 10);
-      expect(binding.cachedPriceMap.keys()).to.empty;
+      expect(multicallStub).callCount(1);
     });
 
     it("should fetch prices using multicall if multicall is configured", async () => {
       multicallStub.resolves([parseEther(MOCK_TOKEN_PRICE1), parseEther(MOCK_ETHER_PRICE)]);
       await binding.bindPrices();
       await delay(interval + 10);
-      const values = binding.cachedPriceMap.values();
-      expect(values.next().value.price.toString()).to.be.eq(parseEther(MOCK_TOKEN_PRICE1).toString());
-      expect(values.next().value.price.toString()).to.be.eq(parseEther(MOCK_ETHER_PRICE).toString());
-      expect(values.next().value.price.toString()).to.be.eq(parseEther(MOCK_TOKEN_PRICE1).toString());
-      expect(values.next().value.price.toString()).to.be.eq(parseEther(MOCK_ETHER_PRICE).toString());
+      expect(multicallStub.callCount).to.be.gt(0);
+      expect(getTokenPriceFromOnChainStub).callCount(0);
     });
 
     it("should fetch prices using Promise if multicall isn't configured", async () => {
-      binding.cachedPriceMap.clear();
       configMock.chainConfig["1337"].multicallAddress = null;
       configMock.chainConfig["1338"].multicallAddress = null;
       getTokenPriceFromOnChainStub.resolves(parseEther(MOCK_TOKEN_PRICE2));
       await binding.bindPrices();
       await delay(interval + 10);
-      const values = binding.cachedPriceMap.values();
-      expect(values.next().value.price.toString()).to.be.eq(parseEther(MOCK_TOKEN_PRICE2).toString());
-      expect(values.next().value.price.toString()).to.be.eq(parseEther(MOCK_TOKEN_PRICE2).toString());
-      expect(values.next().value.price.toString()).to.be.eq(parseEther(MOCK_TOKEN_PRICE2).toString());
-      expect(values.next().value.price.toString()).to.be.eq(parseEther(MOCK_TOKEN_PRICE2).toString());
+      expect(multicallStub).callCount(0);
+      expect(getTokenPriceFromOnChainStub.callCount).to.be.gt(0);
     });
   });
 });