+++ packages/router/test/utils.ts
@@ -56,6 +56,7 @@ export const configMock: NxtpRouterConfig = {
         { assetId: mkAddress("0xc"), chainId: 1337 },
         { assetId: mkAddress("0xf"), chainId: 1338 },
       ],
+      mainnetEquivalent: mkAddress("0xd"),
     },
   ],
   allowedTolerance: 10,
@@ -85,10 +86,15 @@ export const fulfillInputMock: FulfillInput = {
   side: "receiver",
 };
 
+export const mockHashes = {
+  prepareHash: mkBytes32("0xa"),
+};
+
 export const cancelInputMock: CancelInput = {
   amount: variantDataMock.amount,
   expiry: variantDataMock.expiry,
   preparedBlockNumber: variantDataMock.preparedBlockNumber,
+  preparedTransactionHash: mockHashes.prepareHash,
   side: "sender",
 };
 
@@ -105,7 +111,7 @@ export const activeTransactionPrepareMock: ActiveTransaction<"SenderPrepared"> =
     bidSignature: "0xdbc",
     encodedBid: "0xdef",
     encryptedCallData: "0xabc",
-    senderPreparedHash: mkBytes32("0xa"),
+    hashes: { sending: mockHashes },
   },
   status: CrosschainTransactionStatus.SenderPrepared,
 };
@@ -116,7 +122,7 @@ export const activeTransactionFulfillMock: ActiveTransaction<"ReceiverFulfilled"
     callData: "0x",
     relayerFee: "100000",
     signature: "0xabc",
-    receiverFulfilledHash: mkBytes32("0xa"),
+    hashes: { sending: mockHashes, receiving: { ...mockHashes, fulfillHash: mkBytes32("0xb") } },
   },
   status: CrosschainTransactionStatus.ReceiverFulfilled,
 };