+++ CHANGELOG.md
@@ -2,9 +2,24 @@
 
 ## Next Release
 
-- [router] delegate fee calc to chainreader
-- [sdk] integrate chainreader
-- [txservice] introduced provider syncing
+## 0.0.98
+
+- [contracts] Set up Moonriver price oracle
+
+## 0.0.97
+
+- [contracts] Deploy gas price oracle on xdai
+
+## 0.0.96
+
+- [router/sdk] Fix gas price estimation
+
+## 0.0.94
+
+- [router] Delegate fee calc to chainreader
+- [sdk] Integrate chainreader
+- [txservice] Introduced provider syncing
+- [router] Arbitrum `block.number` handling on cancel
 
 ## 0.0.93